db: filter pointer calls by function signatures
The problem is that we would have code like:

	foo->delete(bar);

And there are many things called delete.  Some of them take a struct
ethernet_thing and some take a struct power_management_releated_pointer.
Let's say one of them frees bar but the rest don't.  In the original code
we would complain if it re-used bar.

Now we look at what parameters we are supposed to be passing and if we
are passing ethernet_thing we can ignore the functions that don't match.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>