fix show-parse()'s labels
show_statement() sometimes displays BBs as "L0x0" because
it uses the symbol::bb_target field which is never set when
show_statement() is usefull (before or during linearization).

Fix this by using the address of the underlaying symbol like
done when displaying others BBs (for example for the label at
the beginning of a BB).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>