CSE: use commutativity to identify equivalent instructions
By definition a commutative operation give the same result if its
operands are exchanged.
CSE can use this property when comparing instructions to find
more equivalent instructions and thus eliminate more of them..

Implement this by special-casing commutative ops in insn_compare().

Note: this come for free when all instructions are properly
canonicalized, which is not yet the case.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>