fix missing reload
In dominates() there is the following comment:
  "We don't think two explicitly different symbols ever alias"
but the corresponding test only check if the storage in the
potentialy dominating instruction is a symbol and if so returns 0
meaning that it can't dominate.

But that's without taking in account that a pointer can point
to this symbol and thus that a store via this symbol can dominate
a store via a pointer.

Fix this by changing the test more in accordance to the comment:
return 0 only if the 2 concerned location correspond to 2 *distinct*
symbols (and thus do not return 0 if one of the location is not a symbol).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>