db: prevent slow downs splitting returns apart
When I wrote this, I was trying to prevent writing too much data to the
on disk database because it causes a slow down.  But it turns out that
just splitting the implications apart for isp1362_show_regs() can take an
extremely long time.

My test case is isp1362_show_regs().  This doesn't affect the released code
but it affects some stuff I'm working on.  And it generally seems like
the right thing to do.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>