param_limit: don't set some unchanged limits
This is only supposed to print a limit if it's actually limited.  So say
we pass in 0-3 and it's still 0-3 when we return that doesn't get stored
as a limit.  There are two reasons for this:  1)  Otherwise we would print
a lot of useless information. 2)  If the caller is updated to pass in 4
then Smatch would say that return is impossible.

The bug here is that estate_equiv() looks at relationships between
variables but we only care about the range.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>