show OP_PHI without VOID
As the normal result of branch simplification OP_PHI
can have some VOID in their phi_list, sometimes lots of them.
These list can't be simplified, comacted or so because the
address of the pseudos is used for the pseudo tracking.
But it's annoying that these VOID are displayed by
show_instruction(), it make things sometimes hard to read.

Chnage this by ommiting to display them (when not verbose).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>