fix is_scalar_type()
This helper was introduced to add some warning in conditionals
which must accept arrays and functions as those degenerate as
pointers in this context.

There was some sign of a bug somewhere but it wasn't seen
and as consequence arrays and function type were added to
scalar type (yeah, brown paper bag).

Now that the bug is solved, fix this by removing array and
function type from is_scalar_type().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>