llvm: fix creation of sparsec's tmp files
It seems the intention was to create the tmp files with
the suffixes but the files are create without the suffixes
(while the filename that is used later is with the suffixes).

In all case the non-suffixed file that is created by the
mktemp command is never removed and after a while these
files begin to pile-up. Better to remove them.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>