give default return type in evaluate_call()
Evaluating the return type of a function is a bit complicated.
For user-defined functions this is done inside evaluate_call()
but for builtin function which have an ->evaluate() method it's
this method that must set it. That's quite easy if the return
type is fixed, like 'int' for example, but less so for genearic
functions or when we simply want to keep the given by the
prototype.

To make things more easy for the last case, set by default
the return type as given by the corresponding symbol (the
prototype thus) *before* calling the evaluate method.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>