return an error if too few args
In evaluate_call(), argumenst are evaluated an a diagnostic
is emitted if the number of args is not what is expected.

Good.
However, the processing continues nevertheless.
If too much args were given, this doesn't matter much
but if too few are given we need to check a bit everywhere
for possible NULL args.

Avoid this by returning early an error if there was too few
arguments.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>