finer control over error vs. warnings
Currently, once an error is issued, warnings are no more
issued, only others errors are.

It make sense as once an error is encountered things can
be left in an incoherent state and could cause lots of
useless warnings because of this incoherence.
However, it's also quite annoying as even unrelated warnings
are so silenced and potential bugs stay thus hidden.

Fix this by:
- use a specific flag for this: 'has_error'
- make the distinction between the current phase and some previous
  phases (but for now we only care about parsing vs evaluation)
- if an error has been issued in a previous phase (at parsing)
  warnings are suppressed for the current phase and all future
  phases
- if an error is issued in the current phase (evaluation) the
  flag is reset after each functions/symbols

For example, with this patch, a typing error in function fa()
will not suppress warnings for function fb(), while a parsing
error will still inhibit all further warnings.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>