smdb.py: reformat call_info output
I need a filename to make this work so I added a new smdb command called
"smdb.py call_info <filename> <function>".  It shows the merged PARAM_VALUE
information.

It's still very much garbage, but at least it's a start and adds a bunch
of infrastructure that can be used later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>