flow: remove some buggy unused code
This code has some comments which correctly point out that it's buggy.  It
turns out that it isn't used anyway so it can just be deleted.

Originally this code was used to determine if a struct member was set in
the initializer, but now we just look it up in smatch_extra.c instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>