flow, db: fix how inline function assignments are handled
The problem here is that big_expression_stack is really unreliable and we
should probably just get rid of it.  I think it was putting the inline
statements on the stack so it was returning false instead of true.  This
is used in smatch_function_hooks and it means that we parse the call twice
once as a function call and once as a function assignment call.  It
affected some new code I'm working on where it said that we're calling a
function twice resulting in a double free.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>