balanced: delete this check
The idea here was that it would be an easy way to list functions that need
to be balanced such as lock/unlock, get/put etc.  But in real life and
especially in the kernel that model is too simplistic to be worth while.

We only ever checked local_bh_disable/enable.  It worked badly and
generated a false positive if we enabled first and then disabled.  These
kinds of bugs are already caught by check_locking.c so it's duplicative.

Reported-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>